add facility

+++ Modules/Content/database/migrations/2024_12_03_191148_create_signables_with_details_view.php
@@ -9,6 +9,9 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (config('database.default') == 'sqlite') {
+            return;
+        }
         DB::statement("
             CREATE OR REPLACE VIEW `signables_with_details` AS
             SELECT
@@ -74,6 +77,9 @@ return new class extends Migration
      */
     public function down(): void
     {
+        if (config('database.default') == 'sqlite') {
+            return;
+        }
         DB::statement('DROP VIEW IF EXISTS `signables_with_details`');
     }
 };
